+++ src/main/java/com/binus/thesis/fisheryapp/dto/request/RequestProsesTransaksi.java
@@ -2,6 +2,7 @@ package com.binus.thesis.fisheryapp.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,5 +14,5 @@ public class RequestProsesTransaksi {
 
     private String isNego;
 
-    private int hargaNego;
+    private BigDecimal hargaNego;
 }
