+++ src/main/java/com/binus/thesis/fisheryapp/dto/request/RequestUpdateTransaksi.java
@@ -2,6 +2,7 @@ package com.binus.thesis.fisheryapp.dto.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -11,7 +12,7 @@ public class RequestUpdateTransaksi {
 
     private String idIkan;
 
-    private int jumlah;
+    private BigDecimal jumlah;
 
     private LocalDate tanggalDibutuhkan;
